feat: add unconvert support (#10)
* Continue resolving unconvert errors

* Finish resolving unconvert linting errors

+++ pkgs/amino/wellknown.go
@@ -495,7 +495,7 @@ func DecodeJSONTime(bz []byte, fopts FieldOptions) (t time.Time, err error) {
 
 // NOTE: probably not needed after protobuf v1.25 and after, replace with New().
 func newPBTimestamp(t time.Time) timestamppb.Timestamp {
-	return timestamppb.Timestamp{Seconds: int64(t.Unix()), Nanos: int32(t.Nanosecond())}
+	return timestamppb.Timestamp{Seconds: t.Unix(), Nanos: int32(t.Nanosecond())}
 }
 
 func DecodeJSONPBTimestamp(bz []byte, fopts FieldOptions) (t timestamppb.Timestamp, err error) {
@@ -525,7 +525,7 @@ func newPBDuration(d time.Duration) durationpb.Duration {
 	nanos := d.Nanoseconds()
 	secs := nanos / 1e9
 	nanos -= secs * 1e9
-	return durationpb.Duration{Seconds: int64(secs), Nanos: int32(nanos)}
+	return durationpb.Duration{Seconds: secs, Nanos: int32(nanos)}
 }
 
 func DecodeJSONPBDuration(bz []byte, fopts FieldOptions) (d durationpb.Duration, err error) {
