+++ pkgs/crypto/bip39/bip39.go
@@ -191,7 +191,7 @@ func addChecksum(data []byte) []byte {
 		dataBigInt.Mul(dataBigInt, BigTwo)
 
 		// Set rightmost bit if leftmost checksum bit is set
-		if uint8(firstChecksumByte&(1<<(7-i))) > 0 {
+		if firstChecksumByte&(1<<(7-i)) > 0 {
 			dataBigInt.Or(dataBigInt, BigOne)
 		}
 	}
@@ -215,7 +215,7 @@ func validateEntropyWithChecksumBitSize(bitSize int) error {
 	if (bitSize != 128+4) && (bitSize != 160+5) && (bitSize != 192+6) && (bitSize != 224+7) && (bitSize != 256+8) {
 		return fmt.Errorf(
 			"wrong entropy + checksum size - expected %v, got %v",
-			int((bitSize-bitSize%32)+(bitSize-bitSize%32)/32),
+			(bitSize-bitSize%32)+(bitSize-bitSize%32)/32,
 			bitSize,
 		)
 	}
