+++ pkgs/crypto/keys/client/add.go
@@ -138,8 +138,8 @@ func addApp(cmd *command.Command, args []string, iopts interface{}) error {
 		return nil
 	}
 
-	account := uint32(opts.Account)
-	index := uint32(opts.Index)
+	account := opts.Account
+	index := opts.Index
 
 	// If we're using ledger, only thing we need is the path and the bech32 prefix.
 	if opts.UseLedger {
